chore: more professional structure for rust
