+++ recipes-rust/controller/files/car_controller/Cargo.toml
@@ -1,0 +1,6 @@
+[workspace]
+resolver = "3"
+members = [
+	"controller",
+	"lib/vehicle_control",
+]
