+++ recipes-rust/controller/files/car_controller/LICENSE
@@ -1,0 +1,5 @@
+MIT License
+
+Copyright (c) 2025 Team4_Drivapi
+
+Permission is hereby granted, free of charge, to any person obtaining a copy...
