+++ recipes-rust/controller/files/car_controller/lib/vehicle_control/src/lib.rs
@@ -1,0 +1,13 @@
+pub mod constants;
+pub mod gamepad;
+pub mod motors;
+pub mod perception;
+pub mod stanley;
+pub mod vector;
+
+pub use constants::*;
+pub use gamepad::*;
+pub use motors::*;
+pub use perception::*;
+pub use stanley::*;
+pub use vector::Vector2f;
