+++ recipes-rust/controller/files/car_controller/lib/vehicle_control/src/vector.rs
@@ -1,0 +1,5 @@
+#[derive(Default, Debug, Clone, Copy)]
+pub struct Vector2f {
+    pub x: f64,
+    pub y: f64,
+}
