--- recipes-rust/controller/files/controller/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "controller"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-linux-embedded-hal = "0.4.0"
-memmap2 = "0.9.10"
-pca9685 = "0.1.0"
-pwm-pca9685 = "1.0.0"
-rppal = "0.22.1"
-serialport = "4.9.0"
-socketcan = "3.5.0"
